requirements.txt
# Airflow 프로바이더
apache-airflow-providers-fab
apache-airflow-providers-slack

# 크롤링 관련 패키지
requests
beautifulsoup4
selenium
lxml
webdriver-manager

# 데이터베이스
psycopg2-binary

# 유틸리티
python-dotenv
